pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "crew-jira-connector"
version = "0.1.0"
description = "Bridges Jira issues to Crew Studio jobs"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23", "pytest-cov>=4.0"]

[tool.setuptools]
packages = ["crew_jira_connector", "crew_jira_connector.jira_backends"]

[tool.setuptools.package-dir]
"crew_jira_connector" = "."
"crew_jira_connector.jira_backends" = "jira_backends"
